# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestão de Avarias API",
        "description": "Documentação OpenAPI da aplicação de gestão de tickets, equipamentos, auditoria e relatórios.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Listar utilizadores",
                "description": "Retorna todos os utilizadores (Apenas para Administradores).",
                "operationId": "7a629ec3ab4a321032bfa425c844d09d",
                "responses": {
                    "200": {
                        "description": "Lista de utilizadores"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/inactive": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Inativar utilizador",
                "description": "Inativa um utilizador do sistema.",
                "operationId": "7b6493e78a4c4a9f7bc8e3779329484b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilizador inativado"
                    },
                    "404": {
                        "description": "Utilizador não encontrado"
                    },
                    "422": {
                        "description": "Operação inválida"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/admin/equipment": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Listar equipamentos",
                "description": "Lista equipamentos com a respetiva sala associada.",
                "operationId": "87341b2fc0e90e9b75d9ca3e79d1d8be",
                "responses": {
                    "200": {
                        "description": "Lista de equipamentos"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Criar equipamento",
                "description": "Regista um novo equipamento no sistema.",
                "operationId": "8ffa9ef4e838724e4f4d3219db2bc867",
                "responses": {
                    "201": {
                        "description": "Equipamento criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/admin/equipment/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Eliminar equipamento",
                "description": "Remove fisicamente um equipamento do sistema.",
                "operationId": "cee3a5b031a31fee9e4e0f467310957d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipamento eliminado"
                    },
                    "404": {
                        "description": "Equipamento não encontrado"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Atualizar equipamento",
                "description": "Atualiza os dados de um equipamento existente.",
                "operationId": "1cff2f6cda1a8b435b6bd0704a7ce77e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipamento atualizado"
                    },
                    "404": {
                        "description": "Equipamento não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/admin/rooms": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Listar salas",
                "description": "Lista todas as salas registadas.",
                "operationId": "1f3319f513efbd3420c84ba0caa106e3",
                "responses": {
                    "200": {
                        "description": "Lista de salas"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Criar sala",
                "description": "Cria uma nova sala de trabalho.",
                "operationId": "6d52c96d143252ee80aa9be4f63727c4",
                "responses": {
                    "201": {
                        "description": "Sala criada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/admin/rooms/{id}": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Atualizar sala",
                "description": "Atualiza os detalhes de uma sala.",
                "operationId": "e459df36a3c4723a818b41786061a819",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sala atualizada"
                    },
                    "404": {
                        "description": "Sala não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/admin/rooms/{id}/inactive": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Inativar sala",
                "description": "Inativa uma sala (Gestão lógica / Soft management).",
                "operationId": "c551ae9a067c643951fedef3bab7be05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sala inativada"
                    },
                    "404": {
                        "description": "Sala não encontrada"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/admin/tickets/{id}/approve-budget": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Aprovar orçamento",
                "description": "Aprova um pedido de orçamento associado a um ticket de avaria.",
                "operationId": "5bb9e7d57449a03e864bf9b42b768abc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orçamento aprovado"
                    },
                    "422": {
                        "description": "Pedido inválido"
                    }
                },
                "security": [
                    {
                        "X-Auth-Token": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "X-Auth-Token": {
                "type": "apiKey",
                "name": "X-Auth-Token",
                "in": "header"
            },
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}